perf(encode): store row match positions as u32 to cut peak memory
The Row backend (greedy / row-hash levels) stored absolute match positions
in `row_positions: Vec<usize>` — 8 bytes per slot, the largest match-finder
allocation. On a 1 MiB input at level 5 the compress peak was 24.3 MB vs
5.1 MB for the C reference (4.76x), with this table the dominant term.

Store positions as `u32` (the donor `U32` layout), halving that table. The
absolute cursor resets to 0 per frame and a single Row frame is capped below
`u32::MAX` in `add_data` (a loud assert, mirroring `check_stream_abs_headroom`;
the cursor reset means this only bounds a single ~4 GiB+ frame, far beyond
the decoder's window cap). `ROW_EMPTY_SLOT` becomes `u32::MAX` and every
stored position stays strictly below it, so the sentinel is unambiguous.
Output is byte-identical (same positions, narrower storage).

Part of #341

+++ zstd/src/encoding/match_generator.rs
@@ -112,7 +112,7 @@ pub(crate) const ROW_LOG: usize = 5;
 pub(crate) const ROW_SEARCH_DEPTH: usize = 16;
 pub(crate) const ROW_TARGET_LEN: usize = 48;
 pub(crate) const ROW_TAG_BITS: usize = 8;
-pub(crate) const ROW_EMPTY_SLOT: usize = usize::MAX;
+pub(crate) const ROW_EMPTY_SLOT: u32 = u32::MAX;
 pub(crate) const ROW_HASH_KEY_LEN: usize = 4;
 // HASH_MIX_PRIME now lives in `crate::encoding::fastpath::scalar`; the four
 // per-CPU `hash_mix_u64` variants share it via that module.
